2/4/2019

+++ second/bcl_caffe/utils.py
@@ -47,7 +47,7 @@ def get_paddings_indicator_caffe(actual_num, max_num, axis=0):
     max_num = np.arange(max_num, dtype=int).reshape(max_num_shape)
     # tiled_actual_num: [[3,3,3,3,3], [4,4,4,4,4], [2,2,2,2,2]]
     # tiled_max_num: [[0,1,2,3,4], [0,1,2,3,4], [0,1,2,3,4]]
-    paddings_indicator = actual_num > max_num
+    paddings_indicator = actual_num.astype(int) > max_num
     # paddings_indicator shape: [batch_size, max_num]
     return paddings_indicator
